Revert "healthd: BatteryMonitor: doLogValues: reduce log level to debug"
This reverts commit 52ca2635b9a9995032de9378ef89d7c84a9fd5d4.

Reason for revert: needed for debug and investigations of battery and charging issues. The tooling depends on these lines (and have been for years). You should have these lines only on SOC changes and in the presence of a serious problem. Please do not remove logging without consulting with the BMS SW team first.

Change-Id: Ife0079a7e0c1499b22e843429832b019139c0d89